# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/deposits/fiat": {
      "post": {
        "operationId": "DepositsController_depositFiat",
        "summary": "Deposit fiat (USD/EUR)",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DepositFiatDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Deposit created and balance updated",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "deposit": {
                      "id": "dd0e8400-e29b-41d4-a716-446655440030",
                      "userId": "550e8400-e29b-41d4-a716-446655440000",
                      "fiatCurrency": "USD",
                      "amount": "500",
                      "fee": "2.5",
                      "status": "completed",
                      "createdAt": "2026-04-09T12:00:00.000Z",
                      "completedAt": "2026-04-09T12:00:00.000Z"
                    },
                    "balance": {
                      "id": "aa0e8400-e29b-41d4-a716-446655440010",
                      "asset": "USD",
                      "balance": "10000.50",
                      "balanceAvailable": "10000.50",
                      "balanceLocked": "0"
                    },
                    "transaction": {
                      "id": "bt-2",
                      "type": "deposit",
                      "amount": "500",
                      "balanceBefore": "9500",
                      "balanceAfter": "10000",
                      "refType": "deposit_fiat",
                      "refId": "dd0e8400-e29b-41d4-a716-446655440030",
                      "createdAt": "2026-04-09T12:00:00.000Z"
                    },
                    "meta": {
                      "userId": "550e8400-e29b-41d4-a716-446655440000"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "deposits"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "DepositsController_listFiat",
        "summary": "List fiat deposits",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "description": "ISO date",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "description": "ISO date",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated fiat deposits",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "data": [
                      {
                        "id": "dd0e8400-e29b-41d4-a716-446655440030",
                        "userId": "550e8400-e29b-41d4-a716-446655440000",
                        "fiatCurrency": "USD",
                        "amount": "500",
                        "fee": "2.5",
                        "status": "completed",
                        "createdAt": "2026-04-09T12:00:00.000Z",
                        "completedAt": "2026-04-09T12:00:00.000Z"
                      }
                    ],
                    "total": 1,
                    "meta": {
                      "total": 1,
                      "limit": 20,
                      "offset": 0
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "deposits"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/deposits/fiat/{id}": {
      "get": {
        "operationId": "DepositsController_getFiat",
        "summary": "Get fiat deposit by ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Fiat deposit detail",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "id": "dd0e8400-e29b-41d4-a716-446655440030",
                    "userId": "550e8400-e29b-41d4-a716-446655440000",
                    "fiatCurrency": "USD",
                    "amount": "500",
                    "fee": "2.5",
                    "status": "completed",
                    "createdAt": "2026-04-09T12:00:00.000Z",
                    "completedAt": "2026-04-09T12:00:00.000Z"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "deposits"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/deposits/crypto/address": {
      "post": {
        "operationId": "DepositsController_getCryptoAddress",
        "summary": "Get wallet address for crypto deposit (mock)",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddressRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Mock deposit address for symbol",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "userId": "550e8400-e29b-41d4-a716-446655440000",
                    "symbol": "BTC",
                    "walletAddress": "bc1qtraining000000000000000000000000000000000000",
                    "expiresAt": null
                  }
                }
              }
            }
          }
        },
        "tags": [
          "deposits"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/deposits/crypto": {
      "post": {
        "operationId": "DepositsController_depositCrypto",
        "summary": "Create crypto deposit (credits balance in sandbox)",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DepositCryptoDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Crypto deposit recorded and balance credited",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "deposit": {
                      "id": "ff0e8400-e29b-41d4-a716-446655440050",
                      "userId": "550e8400-e29b-41d4-a716-446655440000",
                      "amount": "0.05",
                      "walletAddress": "bc1qabc",
                      "status": "completed",
                      "createdAt": "2026-04-09T12:00:00.000Z",
                      "symbol": "BTC"
                    },
                    "balance": {
                      "id": "bb0e8400-e29b-41d4-a716-446655440011",
                      "asset": "BTC",
                      "balance": "0.2",
                      "balanceAvailable": "0.2",
                      "balanceLocked": "0"
                    },
                    "transaction": {
                      "id": "bt-3",
                      "type": "deposit",
                      "amount": "0.05",
                      "balanceBefore": "0.15",
                      "balanceAfter": "0.2",
                      "refType": "deposit_crypto",
                      "refId": "ff0e8400-e29b-41d4-a716-446655440050",
                      "createdAt": "2026-04-09T12:00:00.000Z"
                    },
                    "meta": {
                      "userId": "550e8400-e29b-41d4-a716-446655440000"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "deposits"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "DepositsController_listCrypto",
        "summary": "List crypto deposits",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated crypto deposits",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "data": [
                      {
                        "id": "ff0e8400-e29b-41d4-a716-446655440050",
                        "userId": "550e8400-e29b-41d4-a716-446655440000",
                        "amount": "0.05",
                        "walletAddress": "bc1qabc",
                        "status": "completed",
                        "createdAt": "2026-04-09T12:00:00.000Z",
                        "symbol": "BTC"
                      }
                    ],
                    "total": 1,
                    "meta": {
                      "total": 1,
                      "limit": 20,
                      "offset": 0
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "deposits"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/deposits/crypto/{id}": {
      "get": {
        "operationId": "DepositsController_getCrypto",
        "summary": "Get crypto deposit by ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Crypto deposit detail",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "id": "ff0e8400-e29b-41d4-a716-446655440050",
                    "userId": "550e8400-e29b-41d4-a716-446655440000",
                    "amount": "0.05",
                    "walletAddress": "bc1qabc",
                    "status": "completed",
                    "createdAt": "2026-04-09T12:00:00.000Z",
                    "symbol": "BTC"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "deposits"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/admin/users/{userId}/deposits/fiat": {
      "get": {
        "operationId": "AdminDepositsController_listFiatDeposits",
        "summary": "List user fiat deposits (admin only)",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "description": "User ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Max results (1-100)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Skip N results for pagination",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "description": "Start date (ISO 8601)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "description": "End date (ISO 8601)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns fiat deposits",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "data": [
                      {
                        "id": "dd0e8400-e29b-41d4-a716-446655440030",
                        "userId": "550e8400-e29b-41d4-a716-446655440000",
                        "fiatCurrency": "USD",
                        "amount": "500",
                        "fee": "2.5",
                        "status": "completed",
                        "createdAt": "2026-04-09T12:00:00.000Z",
                        "completedAt": "2026-04-09T12:00:00.000Z"
                      }
                    ],
                    "total": 1,
                    "meta": {
                      "total": 1,
                      "limit": 20,
                      "offset": 0
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Admin access required",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 403,
                    "message": "Admin access required",
                    "error": "Forbidden"
                  }
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 404,
                    "message": "User not found",
                    "error": "Not Found"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "admin-deposits"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/admin/users/{userId}/deposits/fiat/{id}": {
      "get": {
        "operationId": "AdminDepositsController_getFiatDeposit",
        "summary": "Get user fiat deposit by ID (admin only)",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "description": "User ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Deposit ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns fiat deposit",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "id": "dd0e8400-e29b-41d4-a716-446655440030",
                    "userId": "550e8400-e29b-41d4-a716-446655440000",
                    "fiatCurrency": "USD",
                    "amount": "500",
                    "fee": "2.5",
                    "status": "completed",
                    "createdAt": "2026-04-09T12:00:00.000Z",
                    "completedAt": "2026-04-09T12:00:00.000Z"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Admin access required",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 403,
                    "message": "Admin access required",
                    "error": "Forbidden"
                  }
                }
              }
            }
          },
          "404": {
            "description": "User or deposit not found",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 404,
                    "message": "User not found",
                    "error": "Not Found"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "admin-deposits"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/admin/users/{userId}/deposits/crypto": {
      "get": {
        "operationId": "AdminDepositsController_listCryptoDeposits",
        "summary": "List user crypto deposits (admin only)",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "description": "User ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Max results (1-100)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Skip N results for pagination",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns crypto deposits",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "data": [
                      {
                        "id": "ff0e8400-e29b-41d4-a716-446655440050",
                        "userId": "550e8400-e29b-41d4-a716-446655440000",
                        "amount": "0.05",
                        "walletAddress": "bc1qabc",
                        "status": "completed",
                        "createdAt": "2026-04-09T12:00:00.000Z",
                        "symbol": "BTC"
                      }
                    ],
                    "total": 1,
                    "meta": {
                      "total": 1,
                      "limit": 20,
                      "offset": 0
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Admin access required",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 403,
                    "message": "Admin access required",
                    "error": "Forbidden"
                  }
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 404,
                    "message": "User not found",
                    "error": "Not Found"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "admin-deposits"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/admin/users/{userId}/deposits/crypto/{id}": {
      "get": {
        "operationId": "AdminDepositsController_getCryptoDeposit",
        "summary": "Get user crypto deposit by ID (admin only)",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "description": "User ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Deposit ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns crypto deposit",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "id": "ff0e8400-e29b-41d4-a716-446655440050",
                    "userId": "550e8400-e29b-41d4-a716-446655440000",
                    "amount": "0.05",
                    "walletAddress": "bc1qabc",
                    "status": "completed",
                    "createdAt": "2026-04-09T12:00:00.000Z",
                    "symbol": "BTC"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Admin access required",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 403,
                    "message": "Admin access required",
                    "error": "Forbidden"
                  }
                }
              }
            }
          },
          "404": {
            "description": "User or deposit not found",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 404,
                    "message": "User not found",
                    "error": "Not Found"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "admin-deposits"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Test Exchange API",
    "description": "Crypto exchange training platform API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "DepositFiatDto": {
        "type": "object",
        "properties": {
          "fiatCurrency": {
            "type": "string",
            "enum": [
              "USD",
              "EUR"
            ]
          },
          "amount": {
            "type": "number",
            "example": 100.5,
            "minimum": 0.01
          },
          "paymentMethodId": {
            "type": "string",
            "description": "Saved payment method ID"
          },
          "paymentMethod": {
            "type": "object",
            "description": "Inline payment (when no saved method)",
            "example": {
              "type": "card",
              "last4": "4242",
              "brand": "visa"
            }
          }
        },
        "required": [
          "fiatCurrency",
          "amount"
        ]
      },
      "AddressRequestDto": {
        "type": "object",
        "properties": {
          "symbol": {
            "type": "string",
            "example": "BTC"
          }
        },
        "required": [
          "symbol"
        ]
      },
      "DepositCryptoDto": {
        "type": "object",
        "properties": {
          "symbol": {
            "type": "string",
            "example": "BTC"
          },
          "amount": {
            "type": "number",
            "example": 0.001,
            "minimum": 0.00001
          },
          "walletAddress": {
            "type": "string",
            "example": "bc1qxy2kgdygjrsqtzq2n0yrf2493p83kkfjhx0wlh"
          },
          "txHash": {
            "type": "string",
            "description": "Blockchain tx hash for confirmation"
          }
        },
        "required": [
          "symbol",
          "amount",
          "walletAddress"
        ]
      },
...
    }
  }
}